api: unify deployment enable evidence gate
